pyproject.toml
[tool.poetry]
name = "wig20-bot"
version = "0.1.0"
description = "Discord bot for WIG20 index"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.3.2"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"
flask = "^3.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
